Decompose MoneyReportHeader into focused subcomponents
Split the ~2,040-line MoneyReportHeader into 5 files:
- MoneyReportHeader.tsx: orchestrator with shared subscriptions
- MoneyReportHeaderContext.tsx: ref-stable context for cross-component triggers
- MoneyReportHeaderPrimaryAction.tsx: switch-based primary action renderer
- MoneyReportHeaderSecondaryActions.tsx: on-demand secondary action config builder
- MoneyReportHeaderModals.tsx: modal dialogs with mount-once trigger registration
